requirements.txt
aiofiles==0.6.0
appnope==0.1.2
autopep8==1.5.4
backcall==0.2.0
bson==0.5.10
certifi==2020.12.5
click==7.1.2
decorator==4.4.2
dnspython==2.0.0
fastapi==0.62.0
gevent==20.9.0
greenlet==0.4.17
h11==0.11.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
jupyter-client==6.1.7
jupyter-core==4.7.0
MarkupSafe==1.1.1
motor==2.3.0
msgpack==1.0.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycodestyle==2.6.0
pydantic==1.7.3
Pygments==2.7.3
pymongo==3.11.2
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2020.4
pyzmq==20.0.0
six==1.15.0
starlette==0.13.6
toml==0.10.2
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
uvicorn==0.13.1
wcwidth==0.2.5
websocket==0.2.1
websockets==8.1
zope.event==4.5.0
zope.interface==5.2.0
